+++ include/unicornlua/context.h
@@ -14,6 +14,7 @@
 
 extern const char * const kContextMetatableName;
 extern const luaL_Reg kContextMetamethods[];
+extern const luaL_Reg kContextInstanceMethods[];
 
 
 class Context {
@@ -39,6 +40,13 @@ private:
 
 int ul_context_save(lua_State *L);
 int ul_context_restore(lua_State *L);
+
+/** Deallocate a context object.
+ *
+ * This function calls `uc_free()` on versions of Unicorn before 1.0.2, and calls
+ * `uc_context_free()` on 1.0.2+. In either case, it will behave as expected.
+ */
+int ul_context_free(lua_State *L);
 
 
 #define get_context_struct(L, index)   \