+++ include/unicornlua/engine.h
@@ -47,6 +47,9 @@ public:
      * there's no reason to keep a context around once it's no longer used inside Lua.
      * Thus, there isn't really any use in allowing a @ref Context to be created in the
      * heap.
+     *
+     * Changed in 1.1.0: This now automatically saves the engine state in the context.
+     * Before, it was necessary to call `update()` on the returned context object.
      */
     Context *create_context_in_lua();
     void restore_from_context(Context *context);